backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
openai==1.54.0
python-dotenv==1.0.0
